+++ errortools/__init__.py
@@ -20,6 +20,7 @@ from _errortools.errno import (
 from _errortools.classes.group import BaseGroup, GroupErrors
 from _errortools.decorator.cache import error_cache
 from _errortools.decorator.deprecated import deprecated, experimental
+from _errortools.decorator.handlers import suppress, convert
 from _errortools.classes.errorcodes import (
     PureBaseException,
     ContextException,
@@ -105,6 +106,8 @@ __all__ = [
     "deprecated",
     "experimental",
     "error_cache",
+    "suppress",
+    "convert",
     "TracebackType",
     "FrameType",
     # classes